Treemacs/Yasnippet/LSP problems
First, company-lsp is no longer supported, so I have taken that out.

Second, when yasnippet is enabled it interferes with treemacs' icons -
both treemacs-all-the-icons and dired-icons.

Third, this may or may not be related to lsp-treemacs, which when
disabled causes the issue to go away. I think. It's a cluster.